Ids should be always longs!

+++ src/main/java/com/p000ison/dev/sqlapi/Database.java
@@ -333,7 +333,7 @@ public abstract class Database {
 
     public abstract boolean existsEntry(TableObject object);
 
-    protected abstract int getLastEntryId(RegisteredTable table);
+    protected abstract long getLastEntryId(RegisteredTable table);
 
     /**
      * Checks whether the class is supported by this database/database engine
